refactor: canPlayType should return native result type
closes #138

BREAKING CHANGE:

`canPlayType()` now returns a native result type via an enum `CanPlayType` instead of a boolean. This is to better align with the native HTMLMediaElement response type.

In addition, a new `shouldPlayType()` method has taken on the existing functionaliy of `canPlayType()` and returns a boolean. It'll return `true` if `canPlayType()` returns `Maybe` or `Probably`.