refactor: back to typescript -_-
Overall I've decided to once and for all move back to TS. The JSDoc
transition was in an effort to avoid compilation processes but it seems
that it's not possible. It's hard to avoid because:

- Newer syntax has to be transpiled to older syntax for even modern
  browsers.
- Have to transpile type declaration files regardless.
- TS adds support for newer features which make life 100x better such as
  decorators.
- Better overall DX coding with TS over JS - much easier to take full
  advantage of it's true power. Some things are really clunky such as
  type casting, generics, declaring globals and so on.

The project naturally drifted back to TS land as things were required so
I decided no need to fake it anymore, it didn't work out. Maybe in the
future this will be possible with better JSDoc support.