fix(player/react): set poster img src fallback to undef, not ''
When loading a Poster image using the React API, I noticed React
presented this warning:

Warning: An empty string ("") was passed to the src attribute.
This may cause the browser to download the whole page again over
the network. To fix this, either do not render the element at all
or pass null to src instead of an empty string.

This warning was triggered even if the Vidstack poster component
was only ever rendered with a string, so seemed to me unworkaroundable.

I traced this back to this line in the Poster component, which
at some point in rendering would fall back to setting `src` to an
empty string.

It appears undefined is always preferable to an empty string here,
so this change would appear to be low risk.

Note that I have picked `undefined` instead of `null` as the warning
advises, since null appears not to typecheck. This also aligns with
the approach in the SliderVideoProvider component.